Split out `test_read_update_write` tests into separate modules
The `test_read_update_write` test can be reused with different inputs
to test many different cases.  Split it out so it can be subclassed
and run alone without dragging in the rest of the device tests.
