+++ hw_device_mgr/mgr/tests/test_mgr_read_update_write.py
@@ -1,0 +1,237 @@
+from .base_test_class import BaseMgrTestClass
+from ...tests.test_device_read_update_write import (
+    TestDeviceRUW as _TestDeviceRUW,
+)
+import re
+import pytest
+import time
+from functools import lru_cache, cached_property
+
+
+class TestHWDeviceMgrRUW(BaseMgrTestClass, _TestDeviceRUW):
+    @pytest.fixture
+    def obj(self, mgr_config, device_config, all_device_data):
+        self.obj = self.device_class()
+        self.obj.init(
+            mgr_config=mgr_config,
+            device_config=device_config,
+            sim_device_data=all_device_data.values(),
+        )
+        yield self.obj
+
+    test_case_key_re = re.compile(r"^d\.([x0-9])\.(.*)$")
+    drive_interface_key_re = re.compile(r"^d([0-9])\.([0-9]+)\.([0-9]+)\.(.*)$")
+
+    @cached_property
+    def drive_addr_to_index_map(self):
+        """Map drive address (& all aliases) to `mgr.devices` list index."""
+        config_class = self.device_base_class.config_class
+        index_map = dict()
+        for i, d in enumerate(self.obj.devices):
+            index_map[d.address] = i
+            for addr in config_class.address_variants(d.address):
+                index_map[addr] = i
+        return index_map
+
+    @cached_property
+    def string_format_kwargs(self):
+        """YAML string `format()` keyword args."""
+        items = enumerate(str(d) for d in self.obj.devices)
+        drive_strs = [item[1] for item in items]
+        return dict(
+            drives=drive_strs,
+            drives3plus=",".join(drive_strs[3:]),
+            all_drives=",".join(drive_strs),
+        )
+
+    @lru_cache
+    def obj_interface_to_test_case_kv(self, interface_key, interface_val):
+        # Translate interface values `Foo {address[6]} bar` to
+        # `Foo (0, 1, 0) bar`
+        val = interface_val
+        if isinstance(val, str):
+            val = val.format(**self.string_format_kwargs)
+
+        # Extract drive address `(0, 16, 0)` from key  `d0.16.0.control_mode`
+        m = self.drive_interface_key_re.match(interface_key)
+        if m is None:
+            return interface_key, val  # Doesn't match fmt. `d.6.control_mode`
+        address = tuple(map(int, m.groups()[:-1]))  # (0, 16, 0)
+        # Get device test case index
+        index = self.drive_addr_to_index_map[address]  # 6
+
+        # Translate test object interface key `d0.16.0.control_mode` to test
+        # case key `d.6.control_mode`
+        key = f"d.{index}.{m.groups()[-1]}"  # "d.6.control_mode"
+
+        return key, val
+
+    def obj_interface_to_test_case(self, data):
+        return dict(
+            self.obj_interface_to_test_case_kv(k, v) for k, v in data.items()
+        )
+
+    def read_update_write_conv_test_data(self):
+        uint16 = self.device_class.data_type_class.uint16
+        for data in (self.test_data, self.ovr_data):
+            for intf, intf_data in data.items():
+                for key in intf_data.keys():
+                    # Test drive keys
+                    match = self.test_case_key_re.match(key)
+                    if not match:
+                        continue  # Only testing drive keys
+                    if match.group(2) in ("status_word", "control_word"):
+                        # Format status_word, control_word for
+                        # readability, e.g. 0x000F
+                        intf_data[key] = uint16(intf_data[key])
+
+    def munge_test_case_data(self, test_case, dst, suffix=""):
+        # Expand d.x.foo -> d.1.foo, etc.; don't clobber
+        # specific case
+        for intf in self.device_class.interface_names:
+            values = test_case.get(intf + suffix, dict())
+            intf_dst = dst.setdefault(intf, dict())
+            for key, val in values.items():
+                match = self.test_case_key_re.match(key)
+                if not match:
+                    intf_dst[key] = val
+                    continue  # Regular attribute, not d.x.foo
+                d_ix, d_key = match.groups()
+                if d_ix == "x":
+                    for i, dev in enumerate(self.obj.devices):
+                        dev_key = f"d.{i}.{d_key}"
+                        if dev_key not in values:
+                            intf_dst[dev_key] = val
+                else:
+                    d_ix = int(d_ix)
+                    dev_key = f"d.{d_ix}.{d_key}"
+                    intf_dst[dev_key] = val
+
+    def munge_interface_data(self, interface):
+        # Translate test case key `d.6.control_mode` to object interface key
+        # `d0.16.control_mode`
+        data_raw = super().munge_interface_data(interface)
+        data = dict()
+        for key, val in data_raw.items():
+            m = self.test_case_key_re.match(key)
+            if m is None:
+                # Doesn't match fmt. `d.6.control_mode`
+                data[key] = val
+            else:
+                index, key = m.groups()  # ("6", "control_mode")
+                dev = self.obj.devices[int(index)]
+                pfx = self.obj.dev_prefix(dev, suffix=dev.slug_separator)
+                dev_key = f"{pfx}{key}"  # "d0.16.control_mode"
+                data[dev_key] = val
+
+        return data
+
+    def override_interface_param(self, interface, ovr_data):
+        for key, val in ovr_data.items():
+            match = self.test_case_key_re.match(key)
+            if match:
+                index, key = match.groups()
+                intf = self.obj.devices[int(index)].interface(interface)
+                intf.update(**{key: val})
+            else:
+                super().override_interface_param(interface, {key: val})
+
+    def check_interface_values(self, interface, indent=4):
+        if interface in {"feedback_out", "command_in"}:
+            # Higher level interfaces to application
+            return self.check_interface_values_higher(interface, indent=indent)
+        else:
+            # Lower level interfaces close to hardware interface
+            return self.check_interface_values_lower(interface, indent=indent)
+
+    def check_interface_values_higher(self, interface, indent=4):
+        # Check feedback_in, command_out, sim interfaces
+        expected_raw = self.test_data[interface]
+        expected = self.obj_interface_to_test_case(expected_raw)
+        actual_raw = self.obj.interface(interface).get()
+        actual = self.obj_interface_to_test_case(actual_raw)
+        # self.print_dict(actual, f"Actual {interface}", indent=2)
+        if super().check_data_values(interface, expected, actual, indent):
+            return True
+        else:
+            print(f"FAILURE at {self.test_desc}")
+            return False
+
+    def split_drive_data(self, data):
+        # The manager is a device, and each of obj.devices is a
+        # device.  The test cases are crammed together with device
+        # keys prefixed with `d.x.`, meaning it applies to all
+        # drives, or prefixed with `d.0.`, meaning it applies to
+        # the first drive, or prefixed with nothing, meaning it
+        # applies to the manager.
+        mgr_data = data.copy()
+        device_data = list([dict() for d in self.obj.devices])
+        for drive_key, val in list(mgr_data.items()):
+            match = self.test_case_key_re.match(drive_key)
+            if not match:
+                continue  # Not a drive_key
+            val = mgr_data.pop(drive_key)
+            index, key = match.groups()
+            device_data[int(index)][key] = val
+        return mgr_data, device_data
+
+    def check_interface_values_lower(self, interface, indent=4):
+        # Check feedback_in, command_out, sim interfaces
+
+        # Prepare expected data
+        expected = self.test_data[interface]
+        mgr_expected, device_expected = self.split_drive_data(expected)
+        for k, v in mgr_expected.items():
+            if isinstance(v, str):
+                mgr_expected[k] = v.format(**self.string_format_kwargs)
+        # self.print_dict(mgr_expected, f"Expected {interface}", indent=2)
+        # for i, d in enumerate(device_expected):
+        #     self.print_dict(d, f"drive_{i}", indent=4)
+
+        # Check actual against expected data
+        passing = True
+        # - Mgr
+        actual_raw = self.obj.interface(interface).get()
+        actual = self.obj_interface_to_test_case(actual_raw)
+        # self.print_dict(actual, f"Actual {interface}", indent=2)
+        passing &= self.check_data_values(
+            interface, mgr_expected, actual, indent=indent
+        )
+
+        # - Devices
+        for i, device in enumerate(self.obj.devices):
+            actual = device.interface(interface).get()
+            # self.print_dict(actual, f"drive_{i}", indent=4)
+            passing &= self.check_data_values(
+                "", device_expected[i], actual, indent=indent, prefix=f"d.{i}."
+            )
+
+        if not passing:
+            print(f"FAILURE at {self.test_desc}")
+        return passing
+
+    def setup_test(self, test_case):
+        super().setup_test(test_case)
+        # Add SV660N feedback_out home_found exception
+        mno = self.missing_not_ok["feedback_out"]
+        for i in range(7):
+            mno.setdefault(f"d.{i}.home_found", set())  # Empty set signifies OK
+
+    def get_feedback_and_check(self):
+        super().get_feedback_and_check()
+        # Asynch param download causes a race condition.  When
+        # feedback_out.param_state == PARAM_STATE_UPDATING, wait for the
+        # param download to complete before the next cycle
+        updating = self.device_base_class.PARAM_STATE_UPDATING
+        test_data = self.test_data["feedback_out"]
+        for i, dev in enumerate(self.obj.devices):
+            if test_data.get(f"d.{i}.param_state", None) != updating:
+                continue
+            # Spin while we wait on the worker
+            timeout, incr = 1, 0.01
+            for i in range(int(timeout / incr)):
+                if dev.config.initialize_params():
+                    break
+                time.sleep(incr)
+            else:
+                print(f"{dev}.initialize_params() never returned True!")
